mobile view dropdown menu (fixes #327) (#421)
