+++ src/components/messageDropdown/MessageDropdown.scss
@@ -10,9 +10,13 @@
   display: flex;
   flex-direction: column;
   z-index: 1000;
+
+  @media screen and (max-width: 425px) {
+    width: 280px;
+  }
 }
 
-.message-dropdown::before {
+/* .message-dropdown::before {
   content: "";
   width: 15px;
   height: 15px;
@@ -22,7 +26,7 @@
   top: -0.8em;
   transform: translate(-50%);
   clip-path: polygon(50% 0%, 0% 100%, 100% 100%);
-}
+} */
 
 .message-dropdown-header {
   display: flex;
