+++ src/components/messengerMain/MessengerMain.scss
@@ -76,7 +76,6 @@
 
   .contact-text {
     font-size: 12px;
-    width: 40%;
     margin-left: 4px;
   }
