+++ src/components/settingsActionModal/SettingsActionModal.scss
@@ -1,3 +1,17 @@
+@mixin errorStyle {
+  background: rgba(58, 58, 58, 0.7);
+  border-radius: 4px;
+  border: 0.5px solid var(--primary-white);
+  width: 100%;
+  height: 48px;
+  margin-top: 20px;
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  padding: 0 1em;
+  color: var(--secondary-red);
+}
+
 .settings-modal-container {
   position: fixed;
   left: 0%;
@@ -22,7 +36,10 @@
   flex-direction: column;
   align-items: flex-start;
   justify-content: space-between;
-  overflow-y: auto;
+
+  @media screen and (max-width: 445px) {
+    width: 96%;
+  }
 
   .settings-modal-title {
     margin-bottom: 20px;
@@ -43,6 +60,8 @@
     display: flex;
     justify-content: space-between;
     align-items: center;
+    flex-direction: row;
+    margin-left: 0;
 
     button {
       width: auto;
@@ -55,35 +74,13 @@
   }
 
   .error {
-    background: rgba(58, 58, 58, 0.7);
-    border-radius: 4px;
-    border: 0.5px solid var(--primary-white);
-    width: 100%;
-    height: 48px;
-    margin-top: 20px;
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    padding: 0 1em;
-    color: var(--secondary-red);
+    @include errorStyle;
     margin-bottom: 2em;
   }
 
   .message {
-    background: rgba(58, 58, 58, 0.7);
-    border-radius: 4px;
-    border: 0.5px solid var(--primary-white);
-    width: 100%;
-    height: 48px;
-    margin-top: 20px;
-    display: flex;
-    align-items: center;
-    justify-content: flex-start;
-    padding: 0 1em;
-    color: var(--primary-color);
+    @include errorStyle;
     margin-bottom: 2em;
   }
 
 }
-
-
